requirements-dev.txt
black==20.8b1
eth-brownie>=1.14.6,<2.0.0
